Add assetHost as config option for CDN's
If the user is choosing to use a CDN to front an S3 bucket then it would be nice to support that. I've added an `assetHost` option to the config object and then enabled that as an option as the host name in the index.js. If the user does not want to use this, they can simply leave it blank and it'll default to the normal option.