style(readme): change to a cleaner usage example
The original method instructed the user to copy the module from the original npm installation path into the project itself. This wasn't so great as it meant we'd then be adding the dependencies to the users ghost project and potentially increasing the project size by quite a bit.

I've changed the method to allow npm to handle the installation of this package, and then we simply require it in and export it as the module. This way the module itself is still controlled by NPM for easy updates in the future, and any deployment / CI system can handle pulling the package and all the dependencies as normal.